fix -Wconversion / -Wunused-variable warnings
- XYZTransform: drop the unused isInt16 local.
- CodeStreamDecompress: make the uint64->double cast explicit in the
  selective-fetch size estimate.
- WaveletReverse: clamp the int16 5/3 DC-shift store in the int domain
  and cast the (in-range) result, instead of clamp<int16_t> which
  narrowed the int sum before clamping (also fixes a latent overflow
  where an out-of-range value wrapped instead of being clamped).

+++ src/lib/core/codestream/decompress/CodeStreamDecompress.cpp
@@ -2175,8 +2175,8 @@ bool CodeStreamDecompress::fetchByTileSelective(
       uint32_t fetchLen;
       if(contiguousProgression && reduce > 0)
       {
-        uint64_t estimated =
-            std::max<uint64_t>(headerFetchSize, (uint64_t)(part->length_ * pixelRatio));
+        uint64_t estimated = std::max<uint64_t>(
+            headerFetchSize, (uint64_t)(static_cast<double>(part->length_) * pixelRatio));
         fetchLen = (uint32_t)std::min<uint64_t>(estimated, part->length_);
       }
       else
